tests/wpt/tests/css/css-animations/parsing/animation-delay-shorthand-computed.html

<!DOCTYPE html>
<title>animation-delay shorthand (computed values)</title>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#propdef-animation-delay">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>
<script>
test_computed_value("animation-delay", "1s");
test_computed_value("animation-delay", "-1s");
test_computed_value("animation-delay", "1s 2s");
test_computed_value("animation-delay", "1s, 2s");
test_computed_value("animation-delay", "1s 2s, 3s");
test_computed_value("animation-delay", "1s, 2s 3s");
test_computed_value("animation-delay", "1s, 2s, 3s");
</script>
